fix(java): batch-8 - DownloaderFeedback signatures + restore Turtle log layout
Run 35488522430 (be4d6cb) next javac layer:

- Forge/FabricLike/Cleanroom DownloadTask: updateProgress(long,long) ->
  (int,int) to match the merged Tools.DownloaderFeedback interface (the
  old-Turtle interface used long; Kotlin impls already use Int).
- Restore old Turtle's view_logger layout as view_logger_turtle.xml with
  its own binding: com.endiq.LoggerView (game screen + Java GUI launcher)
  needs the Turtle ids (toggle_log, share_log, log_filter, ...) including
  the share button and log filter that Amethyst's shared view_logger.xml
  lacks. Amethyst's com.kdt.LoggerView keeps the original layout
  (activity_basemain) - both log views now compile with their own UIs.

Preemptive sweeps, all clean: every Java ViewBinding reference (single-
and multi-binding files) resolves against its layout's ids.

+++ app_pojavlauncher/src/main/java/com/endiq/turtlelauncher/feature/mod/modloader/CleanroomDownloadTask.java
@@ -34,7 +34,7 @@ public class CleanroomDownloadTask implements InstallTask, Tools.DownloaderFeedb
     }
 
     @Override
-    public void updateProgress(long curr, long max) {
+    public void updateProgress(int curr, int max) {
         int progress100 = (int) (((float) curr / (float) max) * 100f);
         ProgressKeeper.submitProgress(ProgressLayout.INSTALL_RESOURCE, progress100, R.string.mod_download_progress, mVersion);
     }
